Refactor the MarathonBuilder functionality, so that the update/retry functionality exists in a single place.
